app/requirements/base.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.0
django-extensions==3.0.3
django-log-request-id==1.5.0
djangorestframework==3.11.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
